feat: Add ngx-charts library for charting capabilities
This commit adds the "@swimlane/ngx-charts" library to the project, enabling the use of advanced charting capabilities. This library will be used to create new charts for service-based selections in the tracing table.

Refactor AuthService role to "coordinator"

The AuthService role has been updated from "admin" to "coordinator" to align with the role naming convention used in the application.

Update IPrivacyRequestStats interface properties

The properties in the IPrivacyRequestStats interface have been renamed to better reflect their meaning. The "inSLA" property has been changed to "meetsSLA", the "nearingSLAInAWeek" property has been changed to "nearingSLA", and the "breached" property has been changed to "exceededSLA".

Update TracingComponent role check

The role check in the TracingComponent has been updated to use the "isCoordinator" method instead of the "isAdmin" method. This change ensures that the correct role is checked when determining certain functionalities in the component.

Update AppComponent role to "coordinator"

The role in the AppComponent has been updated from "admin" to "coordinator" to align with the role naming convention used in the application.

Update AppComponent role dropdown options

The role dropdown options in the AppComponent have been updated to display "Coordinator" instead of "Administrator" for the "coordinator" role.

Add enums for chart helper

This commit adds enums for chart helper functionalities, including SLAChartLabels, SeriesNames, EChartType, and SLAStatusColors. These enums will be used for chart labeling and styling in the application.

Update DashboardModule imports

The DashboardModule imports have been updated to include the NgxChartsModule from the "@swimlane/ngx-charts" library. This module will be used to import the necessary charting components for the dashboard module.

Update TracingComponent template

The TracingComponent template has been updated to use the "isCoordinator" method instead of the "isAdmin" method for role-based conditional rendering. This change ensures that the correct functionalities are displayed based on the user's role.